requirements.txt
pyyaml
pandas
xmltodict
requests
lxml
sklearn